requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.5.0
dill==0.4.0
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.6.1
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.4.2
kaggle==1.7.4.2
kagglehub==0.3.11
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pointnet2_ops @ file:///home/admin/don/PCT/pointnet2_ops_lib
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.15
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
